third_party/blink/web_tests/external/wpt/fetch/private-network-access/shared-worker-fetch.tentative.https.window-expected.txt

This is a testharness.js-based test.
[PASS] local to local: success.
[FAIL] private to local: failure.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[PASS] private to local: success.
[PASS] private to private: success.
[FAIL] public to local: failed preflight.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[PASS] public to local: success.
[FAIL] public to private: failed preflight.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[PASS] public to private: success.
[PASS] public to public: success.
[FAIL] treat-as-public to local: failed preflight.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[PASS] treat-as-public to local: success.
[PASS] treat-as-public to local (same-origin): no preflight required.
[FAIL] treat-as-public to private: failed preflight.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[PASS] treat-as-public to private: success.
[PASS] treat-as-public to public: success.
Harness: the test ran to completion.
